Extract bookkeeping from api
